# PanicAPI.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PanicAPI (CryptoPanic Gateway)",
    "version": "1.0.0",
    "description": "Render üzerinde çalışan CryptoPanic proxy. Header `X-API-Key` ile korunur; sunucu tarafında CRYPTOPANIC_TOKEN query'e eklenir."
  },
  "servers": [
    {
      "url": "https://panicapi.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "GatewayKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Gateway anahtarınız (GATEWAY_API_KEY)"
      }
    },
    "schemas": {}
  },
  "security": [
    {
      "GatewayKey": []
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "panicapi_health_get",
        "summary": "Healthcheck",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/cryptopanic/posts": {
      "get": {
        "operationId": "panicapi_news_cryptopanic_posts_get",
        "summary": "CryptoPanic Posts (tüm filtreler)",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "news",
                "media"
              ]
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "rising",
                "hot",
                "bullish",
                "bearish",
                "important",
                "saved",
                "lol"
              ]
            }
          },
          {
            "name": "currencies",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "regions",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "en,de,nl,es,fr,it,pt,ru,tr,ar,cn,jp,ko"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "public",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "following",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "metadata",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "PRO: description,image"
          },
          {
            "name": "approved",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Startup/Enterprise: source.url"
          },
          {
            "name": "panic_score",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "rss"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK (JSON veya RSS/XML)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server misconfig or upstream error"
          }
        }
      }
    },
    "/v1/news/cryptopanic/portfolio": {
      "get": {
        "operationId": "panicapi_news_cryptopanic_portfolio_get",
        "summary": "CryptoPanic Portfolio (private)",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server misconfig or upstream error"
          }
        }
      }
    }
  }
}